requirements.txt
gunicorn == 20.1.0
Django==3.2.8
whitenoise == 6.0.0
dj-database-url == 0.5.0
Pillow == 9.0.1
psycopg2==2.9.2
django-jazzmin==2.4.9
django-registration==3.1.2
django-tz-detect==0.4.0
cloudinary==1.29.0
django-cloudinary-storage==0.3.0
django-heroku==0.3.1
django-ckeditor==6.4.0
